Organize GitHub helpers by operation
