+++ tools/common/github/client.go
@@ -1,0 +1,51 @@
+package github
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"os"
+	"os/exec"
+	"strings"
+	"time"
+)
+
+const defaultTimeout = 30 * time.Second
+
+func getJSON(ctx context.Context, path string, out any) error {
+	output, err := commandOutput(ctx, defaultTimeout, "api", path)
+	if err != nil {
+		return err
+	}
+	if err := json.Unmarshal(output, out); err != nil {
+		return fmt.Errorf("failed to parse GitHub response for %s: %w", path, err)
+	}
+	return nil
+}
+
+func commandOutput(ctx context.Context, timeout time.Duration, args ...string) ([]byte, error) {
+	ctxTimeout, cancel := context.WithTimeout(ctx, timeout)
+	defer cancel()
+
+	cmd := exec.CommandContext(ctxTimeout, "gh", args...)
+	output, err := cmd.Output()
+	if err != nil {
+		if exitErr, ok := err.(*exec.ExitError); ok {
+			return nil, fmt.Errorf("gh %s failed: %w\nstderr: %s", strings.Join(args, " "), err, string(exitErr.Stderr))
+		}
+		return nil, fmt.Errorf("failed to execute gh %s: %w", strings.Join(args, " "), err)
+	}
+	return output, nil
+}
+
+func commandRun(ctx context.Context, timeout time.Duration, args ...string) error {
+	ctxTimeout, cancel := context.WithTimeout(ctx, timeout)
+	defer cancel()
+
+	cmd := exec.CommandContext(ctxTimeout, "gh", args...)
+	cmd.Stderr = os.Stderr
+	if err := cmd.Run(); err != nil {
+		return fmt.Errorf("gh %s failed: %w", strings.Join(args, " "), err)
+	}
+	return nil
+}
